shared: use size_t for strbuf
The unsigned datatype could overflow on 64 bit systems with sufficiently
large index files. Switch to size_t just to be safe.

Since the API of strbuf is not exported, this has no side-effect for
libkmod users.

+++ shared/strbuf.c
@@ -77,9 +77,9 @@ bool strbuf_pushchar(struct strbuf *buf, char ch)
 	return true;
 }
 
-unsigned strbuf_pushchars(struct strbuf *buf, const char *str)
+size_t strbuf_pushchars(struct strbuf *buf, const char *str)
 {
-	unsigned int len;
+	size_t len;
 
 	assert(str != NULL);
 	assert(buf != NULL);
@@ -101,7 +101,7 @@ void strbuf_popchar(struct strbuf *buf)
 	buf->used--;
 }
 
-void strbuf_popchars(struct strbuf *buf, unsigned n)
+void strbuf_popchars(struct strbuf *buf, size_t n)
 {
 	assert(buf->used >= n);
 	buf->used -= n;
